Rename program again
