+++ test/config-next/remoteva-b.json
@@ -1,6 +1,6 @@
 {
 	"rva": {
-		"userAgent": "boulder-rva-a",
+		"userAgent": "boulder-remoteva-b",
 		"dnsTries": 3,
 		"dnsStaticResolvers": [
 			"10.77.77.77:8343",
