+++ test/config/remoteva-a.json
@@ -1,11 +1,15 @@
 {
 	"rva": {
-		"userAgent": "boulder-rva-b",
+		"userAgent": "boulder-remoteva-a",
+		"debugAddr": ":8211",
 		"dnsTries": 3,
-		"dnsStaticResolvers": [
-			"10.77.77.77:8343",
-			"10.77.77.77:8443"
-		],
+		"dnsProvider": {
+			"dnsAuthority": "consul.service.consul",
+			"srvLookup": {
+				"service": "dns",
+				"domain": "service.consul"
+			}
+		},
 		"dnsTimeout": "1s",
 		"dnsAllowLoopbackAddresses": true,
 		"issuerDomain": "happy-hacker-ca.invalid",
@@ -21,6 +25,7 @@
 		},
 		"grpc": {
 			"maxConnectionAge": "30s",
+			"address": ":9897",
 			"services": {
 				"va.VA": {
 					"clientNames": [
@@ -34,9 +39,7 @@
 				}
 			}
 		},
-		"features": {
-			"DOH": true
-		},
+		"features": {},
 		"accountURIPrefixes": [
 			"http://boulder.service.consul:4000/acme/reg/",
 			"http://boulder.service.consul:4001/acme/acct/"
@@ -44,10 +47,6 @@
 	},
 	"syslog": {
 		"stdoutlevel": 4,
-		"sysloglevel": -1
-	},
-	"openTelemetry": {
-		"endpoint": "bjaeger:4317",
-		"sampleratio": 1
+		"sysloglevel": 4
 	}
 }
