+++ test/startservers.py
@@ -16,21 +16,21 @@ Service = collections.namedtuple('Service', ('name', 'debug_port', 'grpc_port', 
 
 # Keep these ports in sync with consul/config.hcl
 SERVICES = (
-    Service('boulder-remoteva-a',
+    Service('boulder-remoteva-old-a',
         8011, 9397, 'rva.boulder',
         ('./bin/boulder', 'boulder-va', '--config', os.path.join(config_dir, 'va-remote-a.json'), '--addr', ':9397', '--debug-addr', ':8011'),
         None),
-    Service('boulder-remoteva-b',
+    Service('boulder-remoteva-old-b',
         8012, 9498, 'rva.boulder',
         ('./bin/boulder', 'boulder-va', '--config', os.path.join(config_dir, 'va-remote-b.json'), '--addr', ':9498', '--debug-addr', ':8012'),
         None),
-    Service('boulder-rva-a',
+    Service('boulder-remoteva-a',
         8211, 9897, 'rva.boulder',
-        ('./bin/boulder', 'boulder-rva', '--config', os.path.join(config_dir, 'rva-a.json'), '--addr', ':9897', '--debug-addr', ':8211'),
+        ('./bin/boulder', 'boulder-remoteva', '--config', os.path.join(config_dir, 'remoteva-a.json'), '--addr', ':9897', '--debug-addr', ':8211'),
         None),
-    Service('boulder-rva-b',
+    Service('boulder-remoteva-b',
         8212, 9998, 'rva.boulder',
-        ('./bin/boulder', 'boulder-rva', '--config', os.path.join(config_dir, 'rva-b.json'), '--addr', ':9998', '--debug-addr', ':8212'),
+        ('./bin/boulder', 'boulder-remoteva', '--config', os.path.join(config_dir, 'remoteva-b.json'), '--addr', ':9998', '--debug-addr', ':8212'),
         None),
     Service('boulder-sa-1',
         8003, 9395, 'sa.boulder',
@@ -65,11 +65,11 @@ SERVICES = (
     Service('boulder-va-1',
         8004, 9392, 'va.boulder',
         ('./bin/boulder', 'boulder-va', '--config', os.path.join(config_dir, 'va.json'), '--addr', ':9392', '--debug-addr', ':8004'),
-        ('boulder-remoteva-a', 'boulder-remoteva-b', 'boulder-rva-a', 'boulder-rva-b')),
+        ('boulder-remoteva-old-a', 'boulder-remoteva-old-b', 'boulder-remoteva-a', 'boulder-remoteva-b')),
     Service('boulder-va-2',
         8104, 9492, 'va.boulder',
         ('./bin/boulder', 'boulder-va', '--config', os.path.join(config_dir, 'va.json'), '--addr', ':9492', '--debug-addr', ':8104'),
-        ('boulder-remoteva-a', 'boulder-remoteva-b', 'boulder-rva-a', 'boulder-rva-b')),
+        ('boulder-remoteva-old-a', 'boulder-remoteva-old-b', 'boulder-remoteva-a', 'boulder-remoteva-b')),
     Service('boulder-ca-1',
         8001, 9393, 'ca.boulder',
         ('./bin/boulder', 'boulder-ca', '--config', os.path.join(config_dir, 'ca.json'), '--addr', ':9393', '--debug-addr', ':8001'),
